move depreacted test compont to simple_comps
