+++ openmdao/core/tests/test_component.py
@@ -12,7 +12,7 @@ from openmdao.test_suite.components.expl_comp_simple import TestExplCompSimple
 from openmdao.test_suite.components.expl_comp_array import TestExplCompArray
 from openmdao.test_suite.components.impl_comp_simple import TestImplCompSimple
 from openmdao.test_suite.components.impl_comp_array import TestImplCompArray
-from openmdao.test_suite.components.deprecated_comps import TestExplCompDeprecated
+from openmdao.test_suite.components.simple_comps import TestExplCompDeprecated
 from openmdao.devtools.testutil import assert_rel_error
